docs: add sections for few options and TS usage (#38)
